save candidate skill

+++ src/main/java/com/skilltracker/repository/SkillRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SkillRepository extends JpaRepository<Skill, Long> {
-    List<Skill> findBySkillTypeId(Long skillTypeId);
+public interface SkillRepository extends JpaRepository<Skill, Integer> {
+    List<Skill> findBySkillTypeId(int skillTypeId);
 }
